Fixed issue of changes to the DBC Export format.
There appears to have been a change ini the DBC Export format from a REST API.  Either a manifest was added or the order of the files (manifest and notebook contents) was changed.  In any case, changes have been made to make the solution more robust.

We now set the notebook to the zipentry string if the entry is for a notebook (specifically a V1 notebook).  For example:

if (zipEntryJson.has("version") && zipEntryJson.getString("version") == "NotebookV1")